0000799: When doing a split, if old title is also in new requested titles, it should not delete the old title
git-svn-id: svn+ssh://192.168.2.4/svn/vaf-new/trunk@199 923a8852-75ef-4272-858b-d33465a0fa2a

+++ Elite/Vaf/Model/Split.php
@@ -24,22 +24,29 @@ class Elite_Vaf_Model_Split extends Elite_Vaf_Model_Base
         {
             $this->apply($descendant);
         }
-        if(count($descendants))
+        if(count($descendants) && $this->shouldDeleteOldVehicle()  )
         {
             $this->vehicle()->unlink();
         }
     }
     
+    function shouldDeleteOldVehicle()
+    {
+    	$oldVehiclesTitleForCurrentOperatingLevel = $this->vehicle()->getLevel($this->grain)->getTitle();
+    	return !in_array( $oldVehiclesTitleForCurrentOperatingLevel, $this->newTitles );
+    }
+    
     function apply(Elite_Vaf_Model_Vehicle $descendant)
     {
         $titles = $descendant->toTitleArray();
+        
         $levelsToReplace = $this->getSchema()->getPrevLevelsIncluding($this->grain);
         foreach( $levelsToReplace as $levelToReplace )
         {
             $replacementTitle = $this->vehicle()->getLevel($levelToReplace)->getTitle();
             $titles[$levelToReplace] = $replacementTitle;
         }
-        
+    
         foreach($this->newTitles as $replacementTitle)
         {
             $titles[$this->grain] = $replacementTitle;
